Pull common code into BaseOtlpExporter class
