+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpTraceExporterTests.cs
@@ -61,7 +61,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void ToOtlpResourceSpansTest(bool addResource)
+        public void ToOtlpResourceSpansTest(bool includeServiceNameInResource)
         {
             var evenTags = new[] { new KeyValuePair<string, object>("k0", "v0") };
             var oddTags = new[] { new KeyValuePair<string, object>("k1", "v1") };
@@ -75,22 +75,19 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 new OtlpExporterOptions(),
                 new NoopTraceServiceClient());
 
-            if (addResource)
+            var resourceBuilder = ResourceBuilder.CreateEmpty();
+            if (includeServiceNameInResource)
             {
-                exporter.SetResource(
-                    ResourceBuilder.CreateEmpty().AddAttributes(
-                        new List<KeyValuePair<string, object>>
-                        {
-                            new KeyValuePair<string, object>(Resources.ResourceSemanticConventions.AttributeServiceName, "service-name"),
-                            new KeyValuePair<string, object>(Resources.ResourceSemanticConventions.AttributeServiceNamespace, "ns1"),
-                        }).Build());
-            }
-            else
-            {
-                exporter.SetResource(Resources.Resource.Empty);
+                resourceBuilder.AddAttributes(
+                    new List<KeyValuePair<string, object>>
+                    {
+                        new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceName, "service-name"),
+                        new KeyValuePair<string, object>(ResourceSemanticConventions.AttributeServiceNamespace, "ns1"),
+                    });
             }
 
             var builder = Sdk.CreateTracerProviderBuilder()
+                .SetResourceBuilder(resourceBuilder)
                 .AddSource(sources[0].Name)
                 .AddSource(sources[1].Name);
 
@@ -120,7 +117,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 
                 Assert.Single(request.ResourceSpans);
                 var oltpResource = request.ResourceSpans.First().Resource;
-                if (addResource)
+                if (includeServiceNameInResource)
                 {
                     Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == Resources.ResourceSemanticConventions.AttributeServiceName && kvp.Value.StringValue == "service-name");
                     Assert.Contains(oltpResource.Attributes, (kvp) => kvp.Key == Resources.ResourceSemanticConventions.AttributeServiceNamespace && kvp.Value.StringValue == "ns1");
@@ -340,52 +337,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 
             Assert.True(startCalled);
             Assert.True(endCalled);
-        }
-
-        [Theory]
-        [InlineData("key=value", new string[] { "key" }, new string[] { "value" })]
-        [InlineData("key1=value1,key2=value2", new string[] { "key1", "key2" }, new string[] { "value1", "value2" })]
-        [InlineData("key1 = value1, key2=value2 ", new string[] { "key1", "key2" }, new string[] { "value1", "value2" })]
-        [InlineData("key==value", new string[] { "key" }, new string[] { "=value" })]
-        [InlineData("access-token=abc=/123,timeout=1234", new string[] { "access-token", "timeout" }, new string[] { "abc=/123", "1234" })]
-        [InlineData("key1=value1;key2=value2", new string[] { "key1" }, new string[] { "value1;key2=value2" })] // semicolon is not treated as a delimeter (https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md#specifying-headers-via-environment-variables)
-        public void GetMetadataFromHeadersWorksCorrectFormat(string headers, string[] keys, string[] values)
-        {
-            var otlpExporter = new OtlpTraceExporter(new OtlpExporterOptions());
-            var metadata = typeof(OtlpTraceExporter)
-                    .GetMethod("GetMetadataFromHeaders", BindingFlags.NonPublic | BindingFlags.Static)
-                    .Invoke(otlpExporter, new object[] { headers }) as Metadata;
-
-            Assert.Equal(keys.Length, metadata.Count);
-
-            for (int i = 0; i < keys.Length; i++)
-            {
-                Assert.Contains(metadata, entry => entry.Key == keys[i] && entry.Value == values[i]);
-            }
-        }
-
-        [Theory]
-        [InlineData("headers")]
-        [InlineData("key,value")]
-        public void GetMetadataFromHeadersThrowsExceptionOnOnvalidFormat(string headers)
-        {
-            var otlpExporter = new OtlpTraceExporter(new OtlpExporterOptions());
-            try
-            {
-                var metadata = typeof(OtlpTraceExporter)
-                            .GetMethod("GetMetadataFromHeaders", BindingFlags.NonPublic | BindingFlags.Static)
-                            .Invoke(otlpExporter, new object[] { headers }) as Metadata;
-            }
-            catch (Exception ex)
-            {
-                // The exception thrown here is System.Reflection.TargetInvocationException. The exception thrown by the method GetMetadataFromHeaders
-                // is captured in the inner exception
-                Assert.IsType<ArgumentException>(ex.InnerException);
-                Assert.Equal("Headers provided in an invalid format.", ex.InnerException.Message);
-                return;
-            }
-
-            throw new XunitException("GetMetadataFromHeaders did not throw an exception for invalid input headers");
         }
 
         private class NoopTraceServiceClient : OtlpCollector.TraceService.ITraceServiceClient
